# year2024/day02: optimization: avoid concatenating slices.