The previous solution made us of `slices.Concat` to synthesize new reports where one element was skipped. The downside of that is that `slices.Concat` creates new slices every single time, meaning that memory is allocated on the heap and then elements are copied over.

To avoid those allocations, we can implement the skipping of elements by swapping elements around in-place. The improved algorithm is described through comments in the source code, but the runtime improvements are quite substantial. On my laptop:

    $ go test -bench=. -benchmem -count=10 ./year2024/day02 | tee base.txt  # baseline
    $ go test -bench=. -benchmem -count=10 ./year2024/day02 | tee opt.txt   # with optimizations
    $ benchstat base.txt opt.txt
    goos: linux
    goarch: amd64
    pkg: go.saser.se/adventofgo/year2024/day02
    cpu: AMD Ryzen 9 5900HX with Radeon Graphics
    	 │  base.txt   │               opt.txt               │
    	 │   sec/op    │   sec/op     vs base                │
    Part1-16   171.2µ ± 1%   173.2µ ± 1%   +1.15% (p=0.009 n=10)
    Part2-16   332.2µ ± 1%   185.6µ ± 1%  -44.15% (p=0.000 n=10)
    geomean    238.5µ        179.3µ       -24.84%
    
    	 │   base.txt   │               opt.txt                │
    	 │     B/op     │     B/op      vs base                │
    Part1-16   156.8Ki ± 0%   156.8Ki ± 0%        ~ (p=0.487 n=10)
    Part2-16   298.8Ki ± 0%   156.8Ki ± 0%  -47.51% (p=0.000 n=10)
    geomean    216.5Ki        156.8Ki       -27.55%
    
    	 │  base.txt   │                opt.txt                │
    	 │  allocs/op  │  allocs/op   vs base                  │
    Part1-16   2.001k ± 0%   2.001k ± 0%        ~ (p=1.000 n=10) ¹
    Part2-16   5.169k ± 0%   2.002k ± 0%  -61.27% (p=0.000 n=10)
    geomean    3.216k        2.001k       -37.77%
    ¹ all samples are equal